+++ src/test/java/org/osgl/issues/Gh209.java
@@ -26,14 +26,14 @@ import org.osgl.TestBase;
 import org.osgl.util.N;
 import org.osgl.util.S;
 
-public class Gh208 extends TestBase {
-    public static class Foo {
+public class Gh209 extends TestBase {
+    public static class Foo implements Cloneable {
         public String name;
         public int level;
         public boolean flag;
     }
 
-    public static class Bar {
+    public static class Bar implements Cloneable {
         public String name;
         public Foo foo;
     }
@@ -48,12 +48,6 @@ public class Gh208 extends TestBase {
         foo.flag = true;
         source.foo = foo;
 
-        Bar target = $.deepCopy(source).filter("-foo,+foo.name").to(Bar.class);
-        eq(source.name, target.name);
-        Foo targetFoo = target.foo;
-        notNull(targetFoo);
-        eq(foo.name, targetFoo.name);
-        eq(0, targetFoo.level);
-        no(targetFoo.flag);
+        Bar target = $.cloneOf(source);
     }
 }